backend/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
supabase>=2.0.0
icalendar>=5.0.11
python-dateutil>=2.8.2
